Move csv handling to SubmissionService
